Cards (cards v0.2.0)

Provides methods for creating and handling a deck of cards

Functions

contains?(deck, card)

Checks if a deck contains a specific card

create_deck()

Returns a list of strings representing a deck of playing cards

create_hand(hand_size)

Creates a hand of a given size The hand_size argument indicates how many cards should be in the hand.

deal(deck, hand_size)

Divides a deck into a hand and the remainder of the deck. The hand_size argument indicates how many cards should be in the hand. Returns a tuple of {hand, remainder_of_deck}.

Examples

iex> deck = Cards.create_deck()
iex> {hand, deck} = Cards.deal(deck, 5)
iex> hand
["Ace of Hearts", "Two of Hearts", "Three of Hearts", "Four of Hearts", "Five of Hearts"]
iex> deck
["Six of Hearts", "Seven of Hearts", "Eight of Hearts", "Nine of Hearts", "Ten of Hearts", "Jack of Hearts", "Queen of Hearts", "King of Hearts", "Ace of Diamonds", "Two of Diamonds", "Three of Diamonds", "Four of Diamonds", "Five of Diamonds", "Six of Diamonds", "Seven of Diamonds", "Eight of Diamonds", "Nine of Diamonds", "Ten of Diamonds", "Jack of Diamonds", "Queen of Diamonds", "King of Diamonds", "Ace of Clubs", "Two of Clubs", "Three of Clubs", "Four of Clubs", "Five of Clubs", "Six of Clubs", "Seven of Clubs", "Eight of Clubs", "Nine of Clubs", "Ten of Clubs", "Jack of Clubs", "Queen of Clubs", "King of Clubs", "Ace of Spades", "Two of Spades", "Three of Spades", "Four of Spades", "Five of Spades", "Six of Spades", "Seven of Spades", "Eight of Spades", "Nine of Spades", "Ten of Spades", "Jack of Spades", "Queen of Spades", "King of Spades"]

load(filename)

Loads a deck from a file

save(deck, filename)

Saves a deck to a file

shuffle(deck)

Shuffles a deck of cards