Cards (cards v0.1.0)

It gives functions to creat, handle and deal with deck of cards

Functions

create_deck()

It creates a deck and retruns card comprhended to values

create_hand(hand_size)

It creates a hand depending on the handsize you pass it

## Examples

  iex> cards = Cards.create_deck
  iex> {hand, _} = Cards.create_hand(1)
  iex> hand
  ["One of Hearts"]

deal(deck, hand_size)

It creates hand depending on the hand_size you pass it

read_file(filename)

It reads the filename you pass it as argument

save(deck, filename)

It saves deck of cards in a file with the name you passed it