Cards (cards v0.1.0)

Provides methods for creating and handling a deck of cards

Functions

create_deck()

create_hand(hand_size)

deal(deck, hand_size)

Divides a deck into a hand. The hand_size argument indicates how many cards should be in the hand.

Examples

iex> import Cards
iex> deck = create_deck
iex> {hand, remainder} = deal(deck, 2)
iex> hand
["Three of Clubs", "Five of Hearts"]

load(filename)

save(deck, filename)