cards v0.1.0 Cards

Provides methods for creating and handling a deck of cards

Functions

contains?(deck, card)

Check if the deck contains the given card. The deck argument is an array created by Cards.create_deck. The card argument is a string. The return of this function is true or false.

Examples

iex> deck = Cards.create_deck
iex> Cards.contains?(deck, "Ace of Diamonds")
true

create_deck()

Creates a deck of cards

create_hand(hand_size)

Creates a deck, shuffle it and then divides a deck into a hand and the remainder of the deck. The hand_size argument indicates how many cards should be in the hand. The return of the function is a tuple containing on the firts index the hand, and on the second the remainder of the deck

deal(deck, hand_size)

Divides a deck into a hand and the remainder of the deck. The deck argument is an array created by Cards.create_deck. The hand_size argument indicates how many cards should be in the hand. The return of the function is a tuple containing on the firts index the hand, and on the second the remainder of the deck

Examples

iex> deck = Cards.create_deck
iex> {hand, second_hand} = Cards.deal(deck, 1)
iex> hand
["Ace of Spades"]

load(filename)

Loads, from the system, the file cointaining the saved deck The filename argument is a string coitaining the file name. The return of the function is an array containing the saved deck.

save(deck, filename)

Save the deck into a file in the project folder. The deck argument is an array created by Cards.create_deck. The filename argument is a string containing the disired file name. The return of this function is the status of saving the file.

Examples

iex> deck = Cards.create_deck
iex> Cards.save(deck, "my_deck")
:ok

shuffle(deck)

Suffle the deck and returns it to the user. The deck argument is an array created by Cards.create_deck