cards v0.1.0 Cards

Provides methods for creating and handling a deck of cards

Functions

contains?(deck, card)

Determines wether a deck contains a given card.

Examples

iex> deck = Cards.create_deck
iex> Cards.contains?(deck, "King of Spades")
false

iex> deck = Cards.create_deck
iex> Cards.contains?(deck, "Four of Hearts")
true

create_deck()

Creates a deck.

Examples

iex> Cards.create_deck()
["Ace of Spades", "Two of Spades", "Three of Spades", "Four of Spades",
"Five of Spades", "Ace of Clubs", "Two of Clubs", "Three of Clubs",
"Four of Clubs", "Five of Clubs", "Ace of Hearts", "Two of Hearts",
"Three of Hearts", "Four of Hearts", "Five of Hearts",
"Ace of Diamonds", "Two of Diamonds", "Three of Diamonds",
"Four of Diamonds", "Five of Diamonds"]

create_hand(hand_size)

Pipelines common used methods and give a hand.

Examples

Cards.create_hand(hand_size)
{ *hand*, *deck* }
{["Ace of Spades", "Two of Heards"], [rest of deck]]}

deal(deck, hand_size)

Divides a deck into a hand and the remainder of the deck. The hand_size argument indicates how many cards should be in the hand.

Examples

iex> deck = Cards.create_deck
iex> {hand, deck} = Cards.deal(deck, 2)
iex> hand
["Ace of Spades", "Two of Spades"]

load(filename)

Loads the deck from a file on the HD.

Examples

Cards.load(filename)
{ deck }

save(deck, filename)

Saves the deck to a file on the HD.

Examples

Cards.save(deck, filename)
{ deck, "myDeck" }

shuffle(deck)

Gets a deck and returns a new randomly shuffled list with the same number of records.

Examples

deck = Cards.shuffle(deck)
Cards.shuffle(deck)
["Five of Spades", "Four of Hearts", "Four of Diamonds", "Ace of Hearts",
"Four of Spades", "Five of Clubs", "Five of Hearts", "Three of Clubs",
"Three of Diamonds", "Three of Hearts", "Four of Clubs", "Ace of Spades",
"Two of Hearts", "Two of Clubs", "Ace of Clubs", "Five of Diamonds",
"Two of Diamonds", "Ace of Diamonds", "Two of Spades", "Three of Spades"]