Cards (cards v0.1.0)

Provides methods for creating and handling a deck of cards

contains?(deck, card)

Determines whether a deck contains a given card.

Examples

iex> deck = Cards.create_deck()
iex> Cards.contains?(deck, "Ace of Spades")
true

create_deck()

Returns a list of string representing a deck of playing cards

Examples

iex(1)> Cards.create_deck
["Ace of Spades", "Two of Spades", "Three of Spades", "Four of Spades",
"Five of Spades", "Six of Spades", "Seven of Spades", "Eight of Spades",
"Nine of Spades", "Ten of Spades", "Ace of Clubs", "Two of Clubs",
"Three of Clubs", "Four of Clubs", "Five of Clubs", "Six of Clubs",
"Seven of Clubs", "Eight of Clubs", "Nine of Clubs", "Ten of Clubs",
"Ace of Hearts", "Two of Hearts","Three of Hearts", "Four of Hearts",
"Five of Hearts", "Six of Hearts", "Seven of Hearts", "Eight of Hearts",
"Nine of Hearts", "Ten of Hearts", "Ace of Diamonds", "Two of Diamonds",
"Three of Diamonds", "Four of Diamonds", "Five of Diamonds", "Six of Diamonds",
"Seven of Diamonds", "Eight of Diamonds", "Nine of Diamonds", "Ten of Diamonds"]

create_hand(hand_size)

deal(deck, hand_size)

Divides a deck into a hand and the remainder of the deck. The hand_size argument indicates how many cards should be in the hand.

Examples

iex> deck = Cards.create_deck()
iex> {hand, deck} = Cards.deal(deck, 1)
iex> hand
["Ace of Spades"]
iex> length(deck)
39

load(filename)

Recieves the filename. Converts the binary values to Lists of Strings, and if file is not found an error message is shown.

Examples

iex> Cards.load("my_deck")
["Ace of Spades", "Two of Spades", "Three of Spades", "Four of Spades",
"Five of Spades", "Six of Spades", "Seven of Spades", "Eight of Spades",
"Nine of Spades", "Ten of Spades", "Ace of Clubs", "Two of Clubs",
"Three of Clubs", "Four of Clubs", "Five of Clubs", "Six of Clubs",
"Seven of Clubs", "Eight of Clubs", "Nine of Clubs", "Ten of Clubs",
"Ace of Hearts", "Two of Hearts", "Three of Hearts", "Four of Hearts",
"Five of Hearts", "Six of Hearts", "Seven of Hearts", "Eight of Hearts",
"Nine of Hearts", "Ten of Hearts", "Ace of Diamonds", "Two of Diamonds",
"Three of Diamonds", "Four of Diamonds", "Five of Diamonds", "Six of Diamonds",
"Seven of Diamonds", "Eight of Diamonds", "Nine of Diamonds",
"Ten of Diamonds"]

save(deck, filename)

Recieves a deck of cards and a filename. Saves the cards on the deck inside the file using the erlang term_to_binary/1

Examples

iex> deck = Cards.create_deck()
iex> Cards.save(deck, "my_deck")
:ok