cards v0.1.0 Cards

Provides methods for creating and handling a deck of cards

Functions

contains?(deck, card)

create_deck()

Returns a list of strings repersenting a deck of playing cards

create_hand(hand_size)

deal(deck, hand_size)

Divides a deck into a hand and the remainder of the deck. The handsize argument indicates how many cards should be in the hand.

hello()

load(filename)

save(deck, filename)

shuffle(deck)

Shuffles the playing cards, given a deck