cards v0.1.0 Cards

Documentation for Cards.

Provides methods for creating and handling a deck of cards.

Functions

contains?(deck, card)

Checks if a string representing a playing card is contained inside a deck of playing cards.

Examples

iex> deck = Cards.create_deck
iex> Cards.contains?(deck, "A of Spades")
true

create_deck()

Returns a list of strings representing a deck of playing cards.

create_hand(hand_size)

deal(deck, hand_size)

Divides a deck into a hand and the remainder of the deck. The hand_size argument indicates how may cards should be in the hand.

Examples

iex> deck = Cards.create_deck
iex> {hand, deck} = Cards.deal(deck, 1)
iex> hand
["A of Spades"]

load(filename)

save(deck, filename)

shuffle(deck)

Shuffles a list of strings representing playing cards