Cards (cards v0.1.0)

Provides methods for creating and handling a deck of cards.

Functions

contains?(deck, card)

create_deck()

Return a list of string with representing a deck of cards.

create_hand(hand_size)

deal(deck, hand_size)

Divides a deck into a gand and the remainder of the deck. The hand_size argument indicates how many cards should be in the hand.

## Examples

  iex> deck = Cards.create_deck
  iex> { hand, deck } = Cards.deal(deck, 1)
  iex> hand
  ["Ace of Spades"]

load(filename)

save(deck, filename)

shuffle(deck)