Cards (cards v0.1.0)

Provide methods for creating and handling deck of cards.

Functions

contains?(deck, card)

Determines whether a deck contains a given card

Examples

iex> deck = Cards.create_deck()
iex> Cards.contains?(deck, "Ace of Spades")
true

create_deck()

Returns a list of strings representing a deck of playing cards.

Examples

iex> deck = Cards.create_deck()
["Ace of Spades", "Two of Spades", "Three of Spades", "Four of Spades",
"Five of Spades", "Ace of Clubs", "Two of Clubs", "Three of Clubs",
"Four of Clubs", "Five of Clubs", "Ace of Hearts", "Two of Hearts",
"Three of Hearts", "Four of Hearts", "Five of Hearts", "Ace of Diamonds",
"Two of Diamonds", "Three of Diamonds", "Four of Diamonds", "Five of Diamonds"]

create_hand(hand_size)

Creates a new deck, shuffles it and deals it with the hand_size specified.

Examples

iex> Cards.create_hand(3)

deal(deck, hand_size)

Divides a deck into a hand & the remainder of the deck. The hand_size argument indicates how many cards should be in the hand.

Examples

iex> deck = Cards.create_deck()
iex> {hand, deck} = Cards.deal(deck, 1)
iex> hand
["Ace of Spades"]

load(filename)

loads our saved file.

Examples

iex> Cards.load("my_deck")

save(deck, filename)

Saves our deck of cards in a file by creating it with the name specified.

Examples

iex> deck = Cards.create_deck()
iex> Cards.save(deck, "my_deck")

shuffle(deck)

Returns us a shuffled deck of cards.

Examples

iex> deck = Cards.create_deck()
iex> Cards.shuffle(deck)