Cards (cards v0.1.0)

Proiveds methods for creating and handling a deck of cards

Functions

contains?(deck, card)

Determines whether a deck contains a given card

Examples

iex> deck = Cards.create_deck
iex> Cards.contains?(deck,"King of Clubs")
true

create_deck()

Return list of cards in a deck

create_hand(hand_size)

deal(deck, hand_size)

Divides a deck into a hand and the remainder of the deck. The hand_size argument indicates how many card should be in hand.

Examples

iex> deck = Cards.create_deck
iex> {hand, deck} = Cards.deal(deck, 1)
iex> hand
["Ace of Spades"]

load(filename)

save(deck, filename)

shuffle(deck)